+++ packages/hardhat-verify/test/unit/index.ts
@@ -61,7 +61,6 @@ describe("verify task", () => {
         },
         contractFQN: "contracts/TestContract.sol:TestContract",
         listNetworks: true,
-        noCompile: true,
       };
       const proccesedArgs = await this.hre.run(TASK_VERIFY_RESOLVE_ARGUMENTS, {
         address,
@@ -70,7 +69,6 @@ describe("verify task", () => {
         libraries: "libraries.js",
         contract: "contracts/TestContract.sol:TestContract",
         listNetworks: true,
-        noCompile: true,
       });
 
       assert.deepEqual(proccesedArgs, expectedArgs);