+++ packages/hardhat-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # hardhat
 
+## 2.15.0
+
+### Minor Changes
+
+- 99995d53b: The sample projects now use the new version of the Toolbox
+
+## 2.14.1
+
+### Patch Changes
+
+- e99498638: Added block numbers for all mainnet hardforks
+
 ## 2.14.0
 
 ### Minor Changes